refactor: move WebSocket API to an op_crate
