+++ cli/Cargo.toml
@@ -23,6 +23,7 @@ path = "./bench/main.rs"
 deno_core = { path = "../core", version = "0.75.0" }
 deno_fetch = { path = "../op_crates/fetch", version = "0.18.0" }
 deno_web = { path = "../op_crates/web", version = "0.26.0" }
+deno_websocket = { path = "../op_crates/websocket", version = "0.1.0" }
 regex = "1.3.9"
 serde = { version = "1.0.116", features = ["derive"] }
 
@@ -84,7 +85,6 @@ nix = "0.19.0"
 chrono = "0.4.15"
 os_pipe = "0.9.2"
 test_util = { path = "../test_util" }
-tokio-tungstenite = "0.11.0"
 tower-test = "0.3.0"
 
 [target.'cfg(unix)'.dev-dependencies]
