+++ runtime/lib.rs
@@ -10,6 +10,7 @@ extern crate log;
 pub use deno_crypto;
 pub use deno_fetch;
 pub use deno_web;
+pub use deno_websocket;
 
 pub mod colors;
 pub mod errors;
